# docs/openapi.yaml
# ============================================================
# Gemelo Cognitivo API — OpenAPI 3.0 Contract
# ============================================================
# This file defines the complete REST API contract
# BEFORE server implementation. Any changes to endpoints,
# schemas, or responses must be reflected here first
# and then in the code.
#
# Last updated: 2026-03-07
# ============================================================

openapi: "3.0.3"

info:
  title: Gemelo Cognitivo API
  description: |
    REST API for the Cognitive Twin Learning System.
    Orchestrates three LLM agents (Profiling, Planning, Explanatory),
    a RAG pipeline with ChromaDB, and an ML trajectory classification model.
  version: "0.1.0"
  contact:
    name: Gemelo Digital Team
  license:
    name: MIT

servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.gemelocognitivo.com
    description: Production (future)

# ============================================================
# SECURITY — Bearer JWT
# ============================================================
security:
  - BearerAuth: []

components:
  # ──────────────────────────────────────────────
  # Security Schemes
  # ──────────────────────────────────────────────
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT token obtained from POST /auth/register or POST /auth/login. Send in header: `Authorization: Bearer <token>`


  # ──────────────────────────────────────────────
  # Reusable Schemas
  # ──────────────────────────────────────────────
  schemas:
    # ── COMMON ─────────────────────────────────
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: "Operation completed successfully"

    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          example: "Resource not found"

    # ── AUTH ───────────────────────────────────
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
          description: "User email address"
        password:
          type: string
          format: password
          minLength: 8
          example: "myPassword123"
          description: "Password (min 8 characters)"
        name:
          type: string
          example: "Maria Garcia"
          description: "User full name"

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
          description: "Registered email address"
        password:
          type: string
          format: password
          example: "myPassword123"
          description: "User password"

    TokenResponse:
      type: object
      required: [access_token, token_type, user_id, name]
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          description: "JWT access token"
        token_type:
          type: string
          default: "bearer"
          example: "bearer"
          description: "Token type"
        user_id:
          type: string
          example: "usr_abc123"
          description: "Authenticated user ID"
        name:
          type: string
          example: "Maria Garcia"
          description: "User name"

    # ── PROFILE ────────────────────────────────
    QuestionnaireAnswers:
      type: object
      required: [user_id, answers]
      properties:
        user_id:
          type: string
          example: "usr_abc123"
          description: "User ID"
        answers:
          type: object
          additionalProperties: true
          example:
            q1: "Python"
            q2: 3
            q3: ["backend", "data"]
          description: >
            Dictionary with questionnaire answers. Key = question ID, Value = user answer.


    CompetencyScore:
      type: object
      required: [competency_id, name, level, score]
      properties:
        competency_id:
          type: string
          example: "comp_prog_001"
        name:
          type: string
          example: "Object-oriented programming"
        level:
          type: string
          enum: [low, medium, high]
          example: "medium"
          description: "Detected level"
        score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          example: 0.65
          description: "Normalized score 0-1"

    CompetencyProfile:
      type: object
      required:
        [user_id, profile_id, competencies, recommended_approach, summary]
      properties:
        user_id:
          type: string
          example: "usr_abc123"
        profile_id:
          type: string
          example: "prof_xyz789"
        competencies:
          type: array
          items:
            $ref: "#/components/schemas/CompetencyScore"
          description: "List of evaluated competencies with their scores"
        recommended_approach:
          type: string
          enum: [GENERALIST, SPECIALIST]
          example: "SPECIALIST"
          description: "Recommended trajectory"
        summary:
          type: string
          example: "The user shows strengths in programming and databases, with areas for improvement in system design."
          description: "Text summary of the profile generated by the agent"

    # ── ROADMAP ────────────────────────────────
    RoadmapGenerateRequest:
      type: object
      required: [user_id, approach]
      properties:
        user_id:
          type: string
          example: "usr_abc123"
          description: "User ID"
        approach:
          type: string
          enum: [GENERALIST, SPECIALIST]
          example: "SPECIALIST"
          description: "Preferred trajectory"

    RoadmapBlock:
      type: object
      required: [block_id, content_id, title, order, completed]
      properties:
        block_id:
          type: string
          example: "blk_001"
        content_id:
          type: string
          example: "course_python_101"
          description: "Course ID in the catalog"
        title:
          type: string
          example: "Introduction to Python"
        order:
          type: integer
          example: 1
        completed:
          type: boolean
          default: false
          example: false

    RoadmapPhase:
      type: object
      required: [phase_order, name, blocks]
      properties:
        phase_order:
          type: integer
          example: 1
        name:
          type: string
          example: "Fundamentals"
        blocks:
          type: array
          items:
            $ref: "#/components/schemas/RoadmapBlock"

    RoadmapResponse:
      type: object
      required: [roadmap_id, user_id, approach, phases, explanation]
      properties:
        roadmap_id:
          type: string
          example: "rm_def456"
        user_id:
          type: string
          example: "usr_abc123"
        approach:
          type: string
          enum: [GENERALIST, SPECIALIST]
          example: "SPECIALIST"
        phases:
          type: array
          items:
            $ref: "#/components/schemas/RoadmapPhase"
        explanation:
          type: string
          example: "This roadmap is recommended because the profile shows high affinity with backend specialization."
          description: "Explanatory Agent's reasoning for why this roadmap is recommended"

    AlternativesResponse:
      type: object
      required: [roadmap_id]
      properties:
        roadmap_id:
          type: string
          example: "rm_def456"
        generalist:
          $ref: "#/components/schemas/RoadmapResponse"
        specialist:
          $ref: "#/components/schemas/RoadmapResponse"

    BlockProgressUpdate:
      type: object
      required: [completed]
      properties:
        completed:
          type: boolean
          default: true
          example: true
          description: "Mark as completed (true) or pending (false)"

    # ── ADMIN ──────────────────────────────────
    AdminMetrics:
      type: object
      properties:
        total_users:
          type: integer
          example: 150
        active_users:
          type: integer
          example: 120
        total_roadmaps:
          type: integer
          example: 95
        active_roadmaps:
          type: integer
          example: 80
        total_courses:
          type: integer
          example: 42
        trajectory_distribution:
          type: object
          properties:
            generalist:
              type: integer
              example: 55
            specialist:
              type: integer
              example: 40
        top_courses:
          type: array
          items:
            type: object
            properties:
              course_id:
                type: string
              title:
                type: string
              times_used:
                type: integer

# ============================================================
# PATHS — API Endpoints
# ============================================================
paths:
  # ────────────────────────────────────────────
  # AUTH — Registration and login
  # ────────────────────────────────────────────

  /auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      description: |
        Register a new user and return a JWT access token.

        **Flow:**
        1. Validate email is not already registered
        2. Hash password (bcrypt)
        3. Persist in `users` table
        4. Generate a JWT access token
      operationId: registerUser
      security: [] # No authentication required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: "Email 'user@example.com' is already registered"
        "422":
          description: Invalid input data (Pydantic validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      tags: [Authentication]
      summary: Log in
      description: |
        Authenticate a user with email and password and return a JWT.

        **Flow:**
        1. Look up user by email
        2. Verify password (bcrypt.checkpw)
        3. Generate a JWT access token
      operationId: loginUser
      security: [] # No authentication required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: "Invalid credentials"

  # ────────────────────────────────────────────
  # PROFILE — Cognitive profile
  # ────────────────────────────────────────────

  /api/profile:
    post:
      tags: [Profile]
      summary: Create cognitive profile
      description: |
        Receive questionnaire answers from the frontend and generate
        the user's cognitive profile via the Profiling Agent.

        **Pipeline:**
        1. Receive `raw_answers` from questionnaire
        2. Send to Profiling Agent (LLM) via orchestrator
        3. Agent analyzes competencies and generates scores
        4. Return `competency_profile` with recommended approach
      operationId: createProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionnaireAnswers"
      responses:
        "201":
          description: Cognitive profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetencyProfile"
        "400":
          description: Invalid questionnaire data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/profile/{user_id}:
    get:
      tags: [Profile]
      summary: Get user cognitive profile
      description: Get the stored cognitive profile for a user.
      operationId: getProfile
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          example: "usr_abc123"
      responses:
        "200":
          description: Profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetencyProfile"
        "404":
          description: User has no cognitive profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ────────────────────────────────────────────
  # ROADMAP — Generation and management
  # ────────────────────────────────────────────

  /api/roadmap:
    post:
      tags: [Roadmap]
      summary: Generate personalized roadmap
      description: |
        Orchestrate the full pipeline to generate a personalized roadmap:

        1. **Profiling Agent** → get the user's `competency_profile`
        2. **RAG (ChromaDB)** → search relevant courses from the catalog
        3. **Planning Agent** → structure phases and blocks
        4. **ML Predict** → classify trajectory (generalist/specialist)
        5. **Explanatory Agent** → generate personalized explanation
      operationId: createRoadmap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoadmapGenerateRequest"
      responses:
        "201":
          description: Roadmap generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadmapResponse"
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/roadmap/{roadmap_id}/alternatives:
    get:
      tags: [Roadmap]
      summary: Get trajectories A and B
      description: |
        Return both trajectories (generalist and specialist) for
        an existing roadmap. Allows the user to compare and choose.
      operationId: getRoadmapAlternatives
      parameters:
        - name: roadmap_id
          in: path
          required: true
          schema:
            type: string
          example: "rm_def456"
      responses:
        "200":
          description: Both trajectories returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlternativesResponse"
        "404":
          description: Roadmap not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/roadmap/{roadmap_id}/block/{block_id}:
    patch:
      tags: [Roadmap]
      summary: Mark block as completed
      description: Mark a roadmap block as completed or pending.
      operationId: updateBlockProgress
      parameters:
        - name: roadmap_id
          in: path
          required: true
          schema:
            type: string
          example: "rm_def456"
        - name: block_id
          in: path
          required: true
          schema:
            type: string
          example: "blk_001"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockProgressUpdate"
      responses:
        "200":
          description: Progress updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Roadmap or block not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ────────────────────────────────────────────
  # ADMIN — Administration panel
  # ────────────────────────────────────────────

  /api/admin/metrics:
    get:
      tags: [Administration]
      summary: Administration panel metrics
      description: |
        Return aggregated metrics:
        - Total users / active
        - Total roadmaps / active
        - Total courses
        - Trajectory A vs B distribution
        - Top 5 most used courses in roadmaps
      operationId: getAdminMetrics
      responses:
        "200":
          description: Metrics returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMetrics"
        "403":
          description: No admin permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"